Updated employee class
